requirements.txt
asgiref==3.2.10
attrs==20.2.0
autobahn==20.7.1
Automat==20.2.0
cffi==1.14.3
channels==2.4.0
constantly==15.1.0
cryptography==3.1.1
daphne==2.5.0
dj-database-url==0.5.0
Django==3.1.1
django-cors-headers==3.5.0
django-heroku==0.3.1
djangorestframework==3.12.1
gunicorn==20.0.4
hyperlink==20.0.1
idna==2.10
incremental==17.5.0
psycopg2==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyHamcrest==2.0.2
pyOpenSSL==19.1.0
pytz==2020.1
service-identity==18.1.0
six==1.15.0
sqlparse==0.3.1
Twisted==20.3.0
txaio==20.4.1
whitenoise==5.2.0
zope.interface==5.1.0
